requirements.txt
certifi==2019.9.11
cffi==1.13.2
Click==7.0
cryptography==2.8
Flask==1.1.1
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.4
pycparser==2.19
PySocks==1.7.1
python-telegram-bot==12.2.0
six==1.13.0
socks==0
tornado==6.0.3
Werkzeug==0.16.0
